requirements.txt
# Bioinformatics and sequence analysis
biopython>=1.83
requests>=2.31.0

# Multiple sequence alignment
clustalo>=0.1.2

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Data handling
pandas>=2.1.0
numpy>=1.26.0
